requirements.txt
einops==0.3.2
fairscale==0.4.13
ipdb==0.13.9
nltk==3.5
numpy==1.23.4
opencv_python==4.5.1.48
opencv_python_headless==4.5.3.56
pandas==1.2.0
Pillow==8.1.0
pydicom==2.1.2
python_dateutil==2.8.2
pytorch_lightning==1.5.10
scikit_image==0.18.1
scikit_learn==0.24.1
scipy==1.6.0
timm==0.4.12
torch==1.12.1
torchmetrics==0.8.2
torchvision==0.13.1
tqdm==4.56.0
transformers==4.18.0
